feat: added global state var to App.js
In this update I expanded upon the global state in App.js to include a list of accounts, the current user and a list of posts. This may be changed in the future to make it simpler, but this is what works with the code right now. In addition, I also changed code in AccountInfo to account for dynamic accounts.